Adds message helper for less tracked changes
