+++ lib/gambit.js
@@ -12,7 +12,7 @@ const platform = 'gambit-slack';
  * @param {Object} data
  * @return {Promise}
  */
- function post(endpoint, query = {}, data = {}) {
+function post(endpoint, query = {}, data = {}) {
   const url = `${config.baseUri}/${endpoint}`;
 
   logger.debug('gambit.post', { url, query, data });
