+++ lib/slack.js
@@ -4,7 +4,8 @@ const logger = require('heroku-logger');
 
 const cache = require('./cache');
 const gambit = require('./gambit');
-const northstar = require('./northstar');
+const { fetchUserByEmail } = require('./northstar');
+const { parseInboundCommand } = require('./message');
 
 /**
  * @param {String} slackUserId
@@ -24,7 +25,7 @@ async function getNorthstarIdBySlackUserId(slackUserId, client) {
 
   const slackRes = await client.users.info({ user: slackUserId });
 
-  const northstarRes = await northstar.fetchUserByEmail(slackRes.user.profile.email);
+  const northstarRes = await fetchUserByEmail(slackRes.user.profile.email);
 
   return cache.set(slackUserId, northstarRes.id);
 }
@@ -109,7 +110,7 @@ async function reply({ message, client }) {
   const { channel } = message;
 
   try {
-    const { command, arg } = parseInboundMessageCommand(message);
+    const { command, arg } = parseInboundCommand(message);
 
     const userId = await getNorthstarIdBySlackUserId(message.user, client);
 
@@ -146,33 +147,6 @@ async function reply({ message, client }) {
     return await sendErrorMessage({ client, channel, error });
   }
 };
-
-/**
- * @param {Object} message
- * @return {Object}
- */
-function parseInboundMessageCommand(message) { 
-  const { text } = message;
-
-  if (text.toLowerCase().trim() === 'photo') {
-    return {
-      command: 'photo',
-      // TODO: Inspect message for attachments instead of hardcoding an image URL.
-      arg: process.env.MEDIA_URL || 'https://user-images.githubusercontent.com/1236811/104386167-6d222700-54e9-11eb-90f1-f7402020e821.JPG',
-    };
-  }
-
-  const params = text.split(' ');
-
-  if (params[0].toLowerCase() === 'broadcast') {
-    return {
-      command: 'broadcast',
-      arg: params[1],
-    };
-  }
-
-  return {};
-}
 
 module.exports = {
   reply,
